renamed addDriveway to CreateDriveway and updated refs
